+++ mitra_validators/src/polls.rs
@@ -4,8 +4,8 @@ use mitra_utils::html::clean_html_all;
 use super::errors::ValidationError;
 
 const POLL_OPTION_COUNT_MIN: usize = 2;
-// https://github.com/mastodon/mastodon/blob/v4.3.8/app/validators/poll_options_validator.rb
-pub const POLL_OPTION_COUNT_MAX: usize = 15;
+// Mastodon can display at least 20
+pub const POLL_OPTION_COUNT_MAX: usize = 20;
 pub const POLL_OPTION_NAME_LENGTH_MAX: usize = 1000;
 
 pub fn clean_poll_option_name(name: &str) -> String {