+++ CHANGELOG.md
@@ -6,10 +6,18 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
 ## [Unreleased]
 
+### Added
+
+- Implemented reverse WebFinger for portable actors.
+
 ### Changed
 
 - Converted `mitra_cli` into library crate.
 - Import object when executing `load-portable-object` command.
+- Increased maximum number of poll options to 20.
+- Improved error message when signed header has invalid value.
+- Add actor ID to "signer not found in cache" error message.
+- Make single HTTP request when searching for profile/post by URL.
 
 ## [4.19.0] - 2026-02-27